+++ toys-core/lib/toys/acceptor.rb
@@ -140,6 +140,19 @@ module Toys
       def convert(str, *extra) # rubocop:disable Lint/UnusedMethodArgument
         str
       end
+
+      ##
+      # Return alternatives for a given non-matching string.
+      #
+      # This method may be called when a match fails. It should return a
+      # (possibly empty) array of alternatives that could be displayed to the
+      # user as "did you mean..."
+      #
+      # The default implementation returns the empty list.
+      #
+      def alternatives(str) # rubocop:disable Lint/UnusedMethodArgument
+        []
+      end
     end
 
     ##
@@ -285,6 +298,14 @@ module Toys
       #
       def convert(_str, elem)
         elem
+      end
+
+      ##
+      # Overrides {Toys::Acceptor::Base#alternatives} to return close matches
+      # from the enum.
+      #
+      def alternatives(str)
+        ::DidYouMean::SpellChecker.new(dictionary: @values.map(&:first)).correct(str)
       end
     end