+++ toys-core/lib/toys/flag_group.rb
@@ -146,7 +146,8 @@ module Toys
         results = []
         flags.each do |flag|
           unless seen.include?(flag.key)
-            results << "Flag \"#{flag.display_name}\" is required."
+            str = "Flag \"#{flag.display_name}\" is required."
+            results << ArgParser::FlagGroupConstraintError.new(str)
           end
         end
         results
@@ -180,10 +181,12 @@ module Toys
           seen_names << flag.display_name if seen.include?(flag.key)
         end
         if seen_names.size > 1
-          ["Exactly one flag out of group \"#{desc}\" is required, but #{seen_names.size}" \
-            " were provided: #{seen_names.inspect}."]
+          str = "Exactly one flag out of group \"#{desc}\" is required, but #{seen_names.size}" \
+                " were provided: #{seen_names.inspect}."
+          [ArgParser::FlagGroupConstraintError.new(str)]
         elsif seen_names.empty?
-          ["Exactly one flag out of group \"#{desc}\" is required, but none were provided."]
+          str = "Exactly one flag out of group \"#{desc}\" is required, but none were provided."
+          [ArgParser::FlagGroupConstraintError.new(str)]
         else
           []
         end
@@ -206,8 +209,9 @@ module Toys
           seen_names << flag.display_name if seen.include?(flag.key)
         end
         if seen_names.size > 1
-          ["At most one flag out of group \"#{desc}\" is required, but #{seen_names.size}" \
-            " were provided: #{seen_names.inspect}."]
+          str = "At most one flag out of group \"#{desc}\" is required, but #{seen_names.size}" \
+                " were provided: #{seen_names.inspect}."
+          [ArgParser::FlagGroupConstraintError.new(str)]
         else
           []
         end
@@ -228,7 +232,8 @@ module Toys
         flags.each do |flag|
           return [] if seen.include?(flag.key)
         end
-        ["At least one flag out of group \"#{desc}\" is required, but none were provided."]
+        str = "At least one flag out of group \"#{desc}\" is required, but none were provided."
+        [ArgParser::FlagGroupConstraintError.new(str)]
       end
 
       ## @private