+++ toys/CHANGELOG.md
@@ -8,6 +8,7 @@ Highlights:
 
 * Changed the license from BSD to MIT to better match how most libraries in the Ruby community are licensed.
 * Tab completion is available for Bash! See the README for instructions on installing it. Tab completion covers tool names, flags, flag values, and positional arguments. Tools can also customize the completion for their own flag and argument values.
+* Toys now integrates with `did_you_mean` to provide suggestions for misspelled tools, flags, and arguments.
 * Tools can now provide their own interrupt handler to respond to user `CTRL-C`. And the default handler no longer displays an unsightly stack trace.
 * A new argument parsing engine, supporting additional features such as optional enforcing that flags appear before positional arguments, as well as a host of fixes, especially around acceptors and optional flag values.
 
@@ -15,6 +16,7 @@ Details:
 
 * CHANGED: Relicensed under the MIT License.
 * ADDED: Tab completion for bash. Args and flags can provide their own completion information.
+* ADDED: The usage error screen displays alternative suggestions when an argument is misspelled.
 * ADDED: Tools can provide an interrupt handler.
 * ADDED: Tools can enforce that flags must be given before positional args. In particular, `toys do` now uses this feature, which eliminates most of the need to use `--` to get flags to work for subtools.
 * ADDED: Function and range based acceptors.
@@ -22,12 +24,15 @@ Details:
 * ADDED: The `:gem_build` template includes an `:install_gem` option.
 * ADDED: The `acceptor` directive takes an optional `type_desc` argument.
 * ADDED: The `accept` directives under flag and positional arg blocks in the DSL can now take blocks and `type_desc` values.
+* ADDED: An `EXTRA_ARGS` context key that provides unmatched positional arguments.
 * ADDED: The Exec util and mixin support specifying a callback for process results.
 * ADDED: The Exec util and mixin provide a way to identify processes by name.
 * CHANGED: Toys now implements its own argument parsing and standard acceptors rather than relying on OptionParser. For the most part, OptionParser behavior is preserved, except in cases where there is clearly a bug.
 * CHANGED: Flags create a short form flag by default if the name has one character.
 * CHANGED: Flags with explicit value-less syntax are no longer given a value even if they specify a default and/or acceptor.
 * CHANGED: Renamed the `TOOL_DEFINITION` context key to `TOOL`, and the corresponding convenience method from `tool_definition` to `tool`.
+* CHANGED: Renamed the `USAGE_ERROR` context key to `USAGE_ERRORS`, and the corresponding convenience method to `usage_errors`. The value is now a (possibly empty) array of `Toys::ArgParser::UsageError` objects.
+* CHANGED: The root tool no longer defines remaining_args.
 * CHANGED: Exec reports failure to start processes in the result object rather than, e.g. raising ENOENT.
 * IMPROVED: Toys no longer displays a stack trace if a tool is interrupted.
 * IMPROVED: Error messages for flag groups are more complete.
@@ -35,6 +40,7 @@ Details:
 * FIXED: Prevented toys-core from being ousted from the load path if a toys file invoked bundler setup.
 * FIXED: Acceptors no longer raise errors when run on missing optional values.
 * FIXED: When reporting errors in toys files, the line number was off by 2.
+* FIXED: The `--usage` help flag now honors `--all` and `--no-recursive`.
 
 Additionally, a significant amount of internal reorganization and cleanup happened in the toys-core gem. See the changelog for toys-core for more details.